feat(chat): pass edit handler to app details modal
