+++ apps/chat/src/components/Marketplace/TabRenderer.tsx
@@ -1,8 +1,14 @@
 import { useCallback, useMemo, useState } from 'react';
 
+import { getFolderIdFromEntityId } from '@/src/utils/app/folders';
+import { isSmallScreen } from '@/src/utils/app/mobile';
+import { ApiUtils } from '@/src/utils/server/api';
+
 import { ApplicationActionType } from '@/src/types/applications';
+import { ShareEntity } from '@/src/types/common';
 import { DialAIEntityModel } from '@/src/types/models';
 import { PublishActions } from '@/src/types/publication';
+import { SharingType } from '@/src/types/share';
 
 import { ApplicationActions } from '@/src/store/application/application.reducers';
 import { useAppDispatch, useAppSelector } from '@/src/store/hooks';
@@ -14,8 +20,10 @@ import {
 
 import { MarketplaceTabs } from '@/src/constants/marketplace';
 
+import { PublishModal } from '@/src/components/Chat/Publish/PublishWizard';
 import { ApplicationDialog } from '@/src/components/Common/ApplicationDialog';
 import { ConfirmDialog } from '@/src/components/Common/ConfirmDialog';
+import ApplicationDetails from '@/src/components/Marketplace/ApplicationDetails/ApplicationDetails';
 import { CardsList } from '@/src/components/Marketplace/CardsList';
 import { MarketplaceBanner } from '@/src/components/Marketplace/MarketplaceBanner';
 import { SearchHeader } from '@/src/components/Marketplace/SearchHeader';
@@ -41,17 +49,10 @@ const deleteConfirmationText = {
 
 interface TabRendererProps {
   entities: DialAIEntityModel[];
-  onCardClick: (entity: DialAIEntityModel) => void;
-  onPublish: (entity: DialAIEntityModel, action: PublishActions) => void;
   isMobile?: boolean;
 }
 
-export const TabRenderer = ({
-  entities,
-  onCardClick,
-  onPublish,
-  isMobile,
-}: TabRendererProps) => {
+export const TabRenderer = ({ entities, isMobile }: TabRendererProps) => {
   const dispatch = useAppDispatch();
 
   const installedModels = useAppSelector(ModelsSelectors.selectInstalledModels);
@@ -65,6 +66,11 @@ export const TabRenderer = ({
     action: DeleteType;
     entity: DialAIEntityModel;
   }>();
+  const [publishModel, setPublishModel] = useState<{
+    entity: ShareEntity;
+    action: PublishActions;
+  }>();
+  const [detailsModel, setDetailsModel] = useState<DialAIEntityModel>();
 
   const handleAddApplication = useCallback(() => {
     setApplicationModel({
@@ -101,23 +107,50 @@ export const TabRenderer = ({
     [deleteModel, installedModels, dispatch],
   );
 
-  const onDelete = useCallback(
+  const handleSetPublishEntity = useCallback(
+    (entity: DialAIEntityModel, action: PublishActions) =>
+      setPublishModel({
+        entity: {
+          name: entity.name,
+          id: ApiUtils.decodeApiUrl(entity.id),
+          folderId: getFolderIdFromEntityId(entity.id),
+        },
+        action,
+      }),
+    [],
+  );
+
+  const handlePublishClose = useCallback(() => setPublishModel(undefined), []);
+
+  const handleDelete = useCallback(
     (entity: DialAIEntityModel) => {
       setDeleteModel({ entity, action: DeleteType.DELETE });
     },
     [setDeleteModel],
   );
 
-  const onRemove = useCallback(
+  const handleRemove = useCallback(
     (entity: DialAIEntityModel) => {
       setDeleteModel({ entity, action: DeleteType.REMOVE });
     },
     [setDeleteModel],
   );
 
-  const onCloseApplicationDialog = useCallback(
+  const handleCardClick = useCallback(
+    (entity: DialAIEntityModel) => {
+      setDetailsModel(entity);
+    },
+    [setDetailsModel],
+  );
+
+  const handleCloseApplicationDialog = useCallback(
     () => setApplicationModel(undefined),
     [setApplicationModel],
+  );
+
+  const handleCloseDetailsDialog = useCallback(
+    () => setDetailsModel(undefined),
+    [setDetailsModel],
   );
 
   const filteredModels = useMemo(() => {
@@ -144,10 +177,10 @@ export const TabRenderer = ({
           selectedTab === MarketplaceTabs.HOME ? 'All applications' : undefined
         }
         entities={filteredModels}
-        onCardClick={onCardClick}
-        onPublish={onPublish}
-        onDelete={onDelete}
-        onRemove={onRemove}
+        onCardClick={handleCardClick}
+        onPublish={handleSetPublishEntity}
+        onDelete={handleDelete}
+        onRemove={handleRemove}
         onEdit={handleEditApplication}
         isMobile={isMobile}
       />
@@ -158,7 +191,7 @@ export const TabRenderer = ({
           isOpen={!!applicationModel}
           isEdit={applicationModel.action === ApplicationActionType.EDIT}
           currentReference={applicationModel.entity?.reference}
-          onClose={onCloseApplicationDialog}
+          onClose={handleCloseApplicationDialog}
         />
       )}
       {!!deleteModel && (
@@ -167,6 +200,24 @@ export const TabRenderer = ({
           {...deleteConfirmationText[deleteModel.action]}
           onClose={handleDeleteClose}
           cancelLabel="Cancel"
+        />
+      )}
+      {detailsModel && (
+        <ApplicationDetails
+          onPublish={handleSetPublishEntity}
+          isMobileView={isMobile ?? isSmallScreen()}
+          entity={detailsModel}
+          onClose={handleCloseDetailsDialog}
+          onEdit={handleEditApplication}
+        />
+      )}
+      {!!(publishModel && publishModel?.entity?.id) && (
+        <PublishModal
+          entity={publishModel.entity}
+          type={SharingType.Application}
+          isOpen={!!publishModel}
+          onClose={handlePublishClose}
+          publishAction={publishModel.action}
         />
       )}
     </>
